Always land on the dashboard after login; status form becomes a card
Two problems, both about what a user sees immediately after signing in.

Login dropped some people on the home page. postLoginPath branched on the
role, and a brand-new account resolves as a plain `user` until an admin
promotes it — so a freshly created sign-in landed on the marketing page with
no hint that a dashboard exists or that access has to be granted. Always go to
the dashboard now and let ProtectedRoute make the call: authorized users see
it, everyone else gets a 403 naming what they're missing. That also removes
the role read from the sign-in path entirely.

The dashboard also still opened on the daily status form, with any granted
cards pushed below it. Extract that form verbatim into DailyStatus and mount
it as a "Daily status" tab on the Timesheet card, so the dashboard is what it
claims to be: the sections you can open. The team tab stays admin-only; daily
status is available to anyone who can open the card.

EmployeeDashboard now carries only the card grid, an empty state that explains
how to get access, and the happiness prompt.

Note the consequence: submitting a daily status now requires the Timesheet
card, so employees who need it must be granted that card.
